[Feat]: 음성 텍스트출력, 모션
