+++ src/index.css
@@ -1,6 +1,6 @@
 body {
   margin: 0;
-  font-family: 'Nanum Gothic', sans-serif;
+  font-family: 'Pretendard', sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
 }
